# pytorrent/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiOk": {
        "properties": {
          "ok": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "Label": {
        "additionalProperties": true,
        "properties": {
          "color": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "profile_id": {
            "type": "integer",
            "default": 1,
            "minimum": 1,
            "description": "Optional rTorrent profile id. Overrides active profile for this request."
          },
          "profile_name": {
            "type": "string",
            "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "Optional per-user API token. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1. When auth is disabled, API endpoints do not require credentials.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearerApiToken": {
        "description": "Optional per-user API token in Authorization: Bearer <token>. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Optional Flask session cookie. Required only when PYTORRENT_AUTH_ENABLE=1 and no API token is used.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    },
    "parameters": {
      "ProfileId": {
        "name": "profile_id",
        "in": "query",
        "required": false,
        "description": "Optional rTorrent profile id for profile-scoped API operations. If omitted, the user active profile is used for backward compatibility; fallback profile id 1 is used only when no active profile exists and the user can access it.",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        }
      },
      "ProfileName": {
        "name": "profile_name",
        "in": "query",
        "required": false,
        "description": "Optional rTorrent profile name for profile-scoped API operations. profile_id has priority when both are supplied.",
        "schema": {
          "type": "string"
        }
      }
    }
  },
  "info": {
    "title": "pyTorrent API",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/labels": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "labels": {
                          "items": {
                            "$ref": "#/components/schemas/Label"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List labels",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Label"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "labels": {
                          "items": {
                            "$ref": "#/components/schemas/Label"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create or update label",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/labels/{label_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "label_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "labels": {
                          "items": {
                            "$ref": "#/components/schemas/Label"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Delete label"
      }
    },
...
    "/api/preferences": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "preferences": {
                          "$ref": "#/components/schemas/Preferences"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get preferences",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "preferences": {
                          "$ref": "#/components/schemas/Preferences"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Save preferences",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/preferences/table-columns/recommended": {
      "post": {
        "description": "Applies the backend-owned recommended torrent table layout, including desktop visibility, mobile visibility, smart mobile filters and column widths.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "preferences": {
                          "$ref": "#/components/schemas/Preferences"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Apply recommended table columns",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/profiles": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "active_profile": {
                          "$ref": "#/components/schemas/Profile"
                        },
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List rTorrent profiles",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "profile": {
                          "$ref": "#/components/schemas/Profile"
                        },
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create rTorrent profile",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/profiles/diagnostics": {
      "get": {
        "description": "Run diagnostics for the active rTorrent profile.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Diagnostics result"
          }
        },
        "summary": "Active profile diagnostics",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/profiles/export": {
      "get": {
        "description": "Export rTorrent profiles.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Profiles export"
          }
        },
        "summary": "Export profiles",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/profiles/import": {
      "post": {
        "description": "Import rTorrent profiles from exported JSON.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "profiles": {
                    "items": {
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Import profiles",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/profiles/test": {
      "post": {
        "description": "Test a submitted SCGI profile payload without saving it.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "is_remote": {
                    "type": "boolean"
                  },
                  "max_parallel_jobs": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "scgi_url": {
                    "type": "string"
                  },
                  "timeout_seconds": {
                    "type": "number"
                  },
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Test rTorrent profile",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/profiles/{profile_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Delete rTorrent profile"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "profile": {
                          "$ref": "#/components/schemas/Profile"
                        },
                        "profiles": {
                          "items": {
                            "$ref": "#/components/schemas/Profile"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Update rTorrent profile"
      }
    },
...
    "/api/profiles/{profile_id}/diagnostics": {
      "get": {
        "description": "Run diagnostics for a saved rTorrent profile.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Diagnostics result"
          }
        },
        "summary": "Profile diagnostics",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
...
  }
}